+++ ess-policy-notifier/ess-policy-notifier-ejb/src_client/main/java/com/avioconsulting/EssPolicyNotifier.java
@@ -1,0 +1,9 @@
+package com.avioconsulting;
+
+import javax.ejb.Remote;
+import java.util.List;
+
+@Remote
+public interface EssPolicyNotifier {
+    List<String> createPolicyAssembly(String hostingApp, String essPackage, String jobName) throws Exception;
+}