+++ src/main/groovy/com/avioconsulting/ess/models/EssClientPolicySubject.groovy
@@ -1,0 +1,30 @@
+package com.avioconsulting.ess.models
+
+import com.avioconsulting.ess.wrappers.MetadataServiceWrapper
+import groovy.transform.Canonical
+import groovy.transform.InheritConstructors
+
+@Canonical
+@InheritConstructors
+class EssClientPolicySubject extends PolicySubject {
+    static final String DEFAULT_ESS_HOST_APP = 'EssNativeHostingApp'
+    String essHostApplicationName
+    JobDefinition jobDefinition
+
+    EssClientPolicySubject(Map map) {
+        super(map << [essHostApplicationName: (map.essHostApplicationName ?: DEFAULT_ESS_HOST_APP)])
+    }
+
+    String getApplicationNameOnly() {
+        essHostApplicationName
+    }
+
+    String getAssembly() {
+        '%WsmPolicy:/' + MetadataServiceWrapper.PACKAGE_NAME_WHEN_CREATED_VIA_EM + this.jobDefinition.name
+    }
+
+    String getSubject() {
+        // the ones created via EM use this
+        'Job-Invoke()'
+    }
+}