+++ src/main/groovy/com/avioconsulting/util/Logger.groovy
@@ -1,0 +1,5 @@
+package com.avioconsulting.util
+
+interface Logger {
+    def info(String message)
+}