+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.2
+
+* Added WSM policy attachment for ESS jobs
+* Added support for 'every minute' schedules
+
 # 1.0.1
 
 * Support monthly schedules